[base] substitution: add more trait functions
It's impossible to use an associated const generic to return plain
`Vector` and `RowMatrix` in trait methods, so we need separate functions
to do these basic calculations.

This might even be for the better, given that it decouples the
presentation from row/column matrix types.

The main problem is excessive function depth.  I have a feeling that
I'll need a lot of `inline(always)`.