[shchurvec] lib: rework how `ShchurVec` works
The initial implementation had an extremely slow `accept`, caused by
necessary copying.  This is a redesign which preserves the core idea of
mask-addressing, but eliminates the copies.

The roles of masking and value status are split into two fields: mask
and edited.  The prior points to the currently active element and the
latter tracks wherever the value was changed.

- `set` toggles the mask, updates the slot pointed to by the mask, and
  marks the field as edited.

- Indexing works as usual, with mask providing the offset.

- `accept` simply clears the edited values, preserving the masks.

- `reject` toggles masks only for edited values, and clears the edits.