[base] Use `anyhow` for error handling
This is considered a bad practice, because it's a library.  But the
problem is, I need an error type which encompasses all of the errors,
including the ones from the downstream crates.  And some of them will
definitely need to use the downstream types, causing a cyclical
dependency.  I can add a `Box<dyn Error>` variant to the enum, but at
this point I'm just making a worse `anyhow::Error`.

So, instead I'll create module-local error types and wrap them with
`anyhow`.  This also allows me to add stack useful call context while
preserving the ability to get the original error with a downcast.