python: improve documentation
I picked Google-style docs to avoid dealing with RST.  The docs are
generated with `pdoc`, which necessitated adding `__all__` in a couple
of places.  It also highlighted non-underscore-prefixed class
attributes.  I'm unsure if I should stick with making them private or
documenting them publicly.  Does it make sense to change a parameter or
a distribution in a prior on the fly?  There are AVMNs, but those have
to be encapsulated.