[b3] state: ditch the `Rng` field
It doesn't make much sense for it to be bundled in state.  The
architecture has moved from state managing everything to MCMC doing
that.  It does get a bit verbose when the RNG has to be passed to all of
the operators, but I think that the decoupling is worth it.
Alternatively it might make sense to switch the `propose` signature to
`(self, rng: Rng)`.

Actually, now that I think about it, maybe I should just get rid of
`State` altogether.  MCMC can manage the parameters and I planned to
remove tree from it in the next commit.  The only thing that's left is
likelihood, but I can store that in MCMC too.