[b3] priors: add a stub Yule model implementation
The simulation needs a coalescent model, otherwise the tree lengths get
scaled to absurdly large numbers.

The current configuration is extremely brittle.  I was trying to fix
runaway branch lengths, so I switched all of the moves to have a
conservative factor factor of 0.9.  But the birth rate scaler must have
a large scaling factor because the optimal value is around 8-9 thousand,
so it must scale out from 1.  The current model settles after around
600k iterations, while swapping the birth rate parameter from 0.1 to
0.75 or the tree scalers from 0.9 to 0.1 breaks it and leaves measly
~200 unique trees after 3M iterations.

All in all, any subtle change can break the entire thing, so it has to
be tested thoroughly.