[b3] operators: optimize Python `weight_of` accesses
FFI is expensive, but I can't cache the weight in node structs because
it's dynamic.  But in these two operators `weight_of` is used inside
iterations and it isn't changed within each step (or is changed, but
only at the end), so it's fine to cache it.

Speaking of it, iteration is also pretty slow.  I wonder if it'd be
beneficial to move the iterator to Python somehow, since all it does is
increment the number.