[b3] likelihood/thread: re-implement
Threaded likelihood is back!  I managed to get it to about 20 minutes
per million on the mushroom benchmark on the server with Vulkan.  And
then it turned out that the CPU calculator can do 13 minutes.  And
that's without threads!

I think Vulkan just isn't made for that kind of workload.  At 6 minutes
you'd have to run almost 3000 shaders per second.  Beagle's custom CUDA
kernels do manage that, so maybe Vulkan is simply not very good for
compute workloads.  What this all means is that I'm abandoning my
emergency plans to roll in transitions into likelihoods because they
don't matter for CPU performance.  And it seems that flamegraph doesn't
count the time spent in the GPU, so my panic about Python being slow was
unfounded.  It's less than 2% for CPU.

The implementation is basically the same as before, except now I use an
Arc to avoid duplicating update data.  I've also changed the `propose`
interface to take vectors at some point, so I changed the order while I
was at it.  I rolled back to slices when I realized that I'd need to
clone at terminal CPU calculators, but decided to keep the order.  It
makes more sense and it seems it was like that some time in the past.