[b3] likelihood/cuda: implement
This is indeed crazy fast.  It diverges on the full primate DNA
benchmark, but not on the mushroom one.  Unsure what the deal here is.
Ah, it's also correct for short primate sequences.  Ah, I forgot to
update length in reject.  Lemme fix that real quick.

And now it's done!  It's 11 minutes on mushrooms.  And likelihood
readback takes up all of the propose space on the flamegraph.  I wonder
how I can optimize that.