[b3] likelihood: rework the API
I realized tonight why my CUDA implementation was still slower.  I'm
calculating real probabilities.  These require two matrix
multiplications.  But that's suboptimal.  Take a single node slide move.
This one node requires recalculation for both branches, since both
lengths have changed.  But all of the parents up the chain only need one
of their child probabilities recalculated.  But in my naive
implementation I recalculate both.

What we can do instead is associate probabilities with edges instead of
nodes.  I called them "projections": a projection of the node likelihood
on the edge.  This way each probability is implicit.  When it's updated,
it has to be recalculated, but sometimes one of its dependent
projections will already be there.

So, I set out to change the API and test it on the CPU implementation.
And for a few hours it was a third slower.  One of the weaknesses of
this method is that we're doing more memory accesses and turns out
storing stuff in non-contiguous chunks really hurts.