[b3] likelihood/cuda: don't use `SkVec`-like structure
`SkVec` works well on the CPU because it avoids expensive backup copies.
But this isn't a concern on the GPU, since `accept` and `reject` run
asynchronously while the Python operators and priors do their thing.
The last timing for 100k on mushrooms was 28 seconds.  Getting rid of
masks speeds it up to 25 seconds, back to the 4 minute 10 seconds result
I projected from my 20k test runs.