[data] tests: create
I don't like having tests in the file, so I think I'll have all of them
in `tests/`.  Now, if only I could get a reliable AI to refactor `stats`
this way and be sure that it didn't mess up anything in the process.
Perhaps coming up with a script would be easier and faster.