[b3] parameter: add accept/reject methods
This supersedes the clunky `backup_params` solution.  I wonder if I even
need these parameters in the first place.  I think that using a native
Python class with an accept/reject protocol might be faster, as that'll
only have one FFI call per parameter per step, as opposed to however
many I'm doing right now.  And because `Parameter` is an enum, it's not
even that type safe from the perspective of the Python code.