[b3] likelihood: add a clock model
For now it doesn't support the full tree, only a single strict rate.
The full support probably needs a conditional `has_updated` check to
avoid recalculating the list for all edges every time.  I could
implement this fully in Python, but I suspect that extracting the edge
rate list to `Vec` would get expensive on larger trees.

I wonder if it'd make sense to add a Python class which is a thin
wrapper around a `Vec`, for more efficiency and faster conversion.