[b3] likelihood/cpu: improve memory layout
This is what BEAST2 does.  It yields more than 2x speedup on the long
primate benchmark.  Which, in hindsight, should've been obvious.  Linear
memory scans are massively better, so instead of jumping back and forth
on each column we should just do those.  They are also cheaper for small
moves which only tweak a few nodes.  The layout I had is terrible, I
wonder how I was even able to get that 80% speedup when the GPU was
idling waiting for memory most of the time.