[b3] likelihood/cuda: perform mass leaf updates separately
This yielded a massive speedup of 2.5 times compared to the old 1.8.
This is actually a simple form of subtree parallelisation because all
leaves are always independent.

Moving forward, it seems that heuristics are the way to go.  I'll need
to setup fine-grained per-move profiling and collect stats on each step
and how much time its calculations take.