[b3] tree: sort internals in `nodes_to_update` by rank
This is in preparation for the CUDA calculator heuristic algorithm
split.  Also, this should make CPU calculator a bit faster, since
there's now more of a chance that nodes go sequentially: sort will fall
back on node order when rank is the same, and node order matches edge
storage order in memory.