[b3] tree: optimize `nodes_to_update`
The last commit was a performance regression.  Turns out sorting is
pretty expensive.  This reuses one allocation and swaps the sorts to
unstable ones, since all of the items are unique anyways.  It still
doesn't fix the problem completely, perhaps at some point in the future
I should sit down and figure out an efficient way to book-keep edited
nodes.