[b3] likelihood/cuda: define launch bounds for `propose` kernel
This speeds things up a bit and finally puts me over the 3x mark on my
benchmark.  But that still doesn't solve the question of how scaling
should be done.  And earlier when I tried to separate cutoff into
`leaves_end` and `internals_start` the codegen worsened and the speed
fell by 20%, unsure why.