[b3] likelihood/cuda: tweaks and a bug fix
I kept trying to separate `leaves_end` and `internals_start` last two
days, but the inference kept getting slower.  That's why: I had a wrong
condition, it should've been `cutoff < 10`, so I wasn't recalculating
leaves most of the time.

Unfortunate, as the 3x mark turned out to be a bust, but at least there
are no blockers to adding the parallel tree algorithm.