[b3] likelihood/cuda: make threads more fine-grained
I was thinking about it completely wrong this whole time.  I thought
that the memory issues came from lack of coalescence, so I was trying to
find ways to pile on more memory onto the threads.  But I'm pretty sure
that the coalescence was just a generic hint from NCU: my patterns were
always coalesced.  The issue was that I was fetching too much memory per
thread.  I was still thinking it terms of vectors, but GPU cores don't
have SIMD instructions, they are SIMD in an of themselves.

This commit switches both `update_leaves` and `propose` to work on
per-allele basis instead of per-site one.  This boosts the benchmark
from 2min 36sec to 2min 24sec, which is pretty decent.

I also removed the `update_internals` function because it added a lot of
complexity but now the 5% advantage was gone too.