[b3] likelihood/cuda: merge `accept` and `reject` CUDA functions
They are doing the same thing: copying from one buffer to another.  I
kept them as two separate functions because I wanted the source buffer
to be `const`.  But now I realized that I can just switch the order of
buffers in parameters and keep one generic version of the function.