[stats] Use `thiserror`
This reduces the amount of boilerplate and I already depend on
`thiserror` in `data`.  For now it's only the Gamma distribution.

I also removed the `Eq` and `Hash` implementations because I'm thinking
about adding the exact invalid values in the errors in the future.  But
I haven't done so yet because it requires a custom `Eq` implementation.
I'm somewhat tempted to just use anyhow in an ad-hoc manner.