[stats] function: stylistic tweaks
I initially started this work with the intent of fixing the `B(1, 1)`
entropy test case.  I tracked down the issue to the `ln_gamma` function,
which returns a small (1e-15) negative number, whereas the true value is
zero.  It turns out this whole thing is non-trivial.  CPython's
implementation works well for a zero, but diverges on larger numbers.
And the precision is also one order of magnitude lower.  The one from
`libm` seems to have a few degenerate cases, both small and large.  I
think the solution will be to test how well each version handles itself
in relative and ULSP terms and pick one depending on `x`.  Now, the only
problem is where do I get "true" values?