[stats] distributions: add Poisson Python PMF and CDF tests
I think there might be a point in setting `num` in `random_` via a flag.
It would enable me to run uber-tests with thousands of cases without
making the default test suite too slow.  The only thing I'm not sure
about is how to handle product parametrization.  Overwriting `num` would
require me to re-import the global setting everywhere, which is verbose.
So a `ratio` argument is probably better.