[math] tolerance: tweak the assert macro
The new version did require the `Comparator` helper struct.  I tried
using variables, but `$opt = $var` gets shadowed, it's just `epsilon =
epsilon`, after all.

Also, `is_close` is suboptimal.  It should be generic.  And `Tolerance`
should also be generic against the output.  Which has to be `T` for
absolute and `f32` or `f64` for relative.

This commit also includes tweaks to the Cauchy tests, which prompted
this change.