[io] newick: rework the parser
The new approach uses Logos instead of `nom`.  I was mostly mucking
around with tokens anyways.  Logos takes care of the low-level parsing
for me.  I do have to write regexes, but, honestly, they are about as
unpleasant as trying to match a float literal in raw code.  It should
also be more performant.  Not that it matters for Newick, but it should
come into play when I move to bigger formats.  Most of them TSV, but
some, like NEXUS, are complex DSLs.