[io] fasta: refactor Rust types to be Sans IO
I moved the IO to the Python implementation for now.  The next step is
to create a unified reader struct for the `io` module and use that
everywhere.

I also added tests, which unearthed a bunch of issues.  The record
constructor requires a leading `>`.  And the Python stub for `DNARecord`
was missing.  And the `DNASeq` comparison was broken.  I'm really ought
to write a full test suite for everything, but the API is still in flux.
I'm not even sure I want to have a separate `DNASeq` type.  Or a
separate `io`-specific `record` type as opposed to something more
generic in `data`.