[b3] likelihood/cuda: implement fine-grained synchronization
This is the most fine-grained version of the sync.  Instead of a warp we
only sync the 4 threads on a single site.  The reason I haven't done
this earlier is because simply replacing `__syncwarp` with `g.sync` was
actually slower.  This version gains performance by doing two things:

- Making the second sync conditional.  This we hit this expensive
  operation much less often.

- Moving `scales` write after the second barrier.  This works even with
  `__syncwarp`, since it minimizes the slow middle section.

All in all, this ekes out another 6% of performance, bringing me to 2
minutes 12 seconds on my reference (sim 500) benchmark.