[b3] builtin/priors: add `Monophyly`
BEAST2's MRCA prior has two modes: monophyletic and a regular one.  The
primate example defines a monophyly on two leaves.  The non-monophyletic
one sets a distribution (or a plain probability decrease) on height.
The latter is actually conceptually easier to implement.  But I'm not
sure how to correctly apply the distribution yet.

Also adds type hints for built-in priors.  I forgot about them.  And
defining a `pyi` woke up pyright, so I had to add all 4 of them at the
same time.