[data] seq: improve parsing speed
So, I decided to benchmark my FASTA parsing.  It turned out to be a
disaster.  87MB/s.  I think the problem is that I'm matching on chars,
not bytes.  And having all that `try_` machinery in the fast path surely
doesn't help either.

This change increases the speed to almost 300MB/s.  It's still extremely
ineffective.  And the solution is very inelegant: I have this weird
undocumented `from_byte` method in the `Character` trait now.

The very long-term solution is probably to specialize the string parsing
function and do SIMD.  But for now I need to nail down the basics.