[data] Tweak the `Character` trait
Every time I write "`Character` trait" in commits it gives me a chuckle.

The previous implementation was a weird one.  I wanted to reuse the
built-in traits.  But those turned out to too inefficient.  The enum
gets bloated to 8 bytes and there's a formatting slow path.  This new
definition should be cleaner and more efficient.  Getting rid of the
ugly casts in `b3` immediately validated this change in my eyes.