[data] newick: tweak the Python API
Removed the `Node` type for now.  It was an ancient holdover from back
when I thought I needed an `Arc` for all Python interface classes.

The interface here is tricky.  I first switched `Node` to a thin wrapper
over `NodeIndex`, like I did for `b3`'s tree.  But `b3` nodes only have
one value: weight.  Newick's node can have attributes.  So, this means
that I might need two classes.  Or two functions for getting either a
name or an attribute from the tree, but that's not exactly ergonomic.

The other option is one I investigated for `b3` before dropping it for
performance reasons.  Each Python node class can made a pair of the
index and a reference to the tree.  It duplicates APIs, but since I
probably won't be using Newick in Rust code, it might make more sense
here.

In fact, a lot of things become easier once performance is no longer a
hard requirement.  Python version of `edges_of` can simply collect the
edges.  Although, it might cause correctness issues if the graph is
edited in between iterations.

It might be wiser to simply drop most of the Newick internals and only
leave thin back and forth conversion shims.

But in either case I do need some kind of flexible and not awfully slow
tree data structure.  But designing one will require a deep dive in how
tree processing is usually done in existing projects and pipelines.

So, with a Rust-to-string serializer in hand, I think I'll shelve Newick
and flexible trees for now.