[data] seq: switch to concrete types backed by `Bytes` and `BytesMut`
I realized that I was beginning to re-implement a homegrown version of
the `bytes` crate, so I switched directly to it.  These stucts are
pretty big (4 words), but I don't think it should be too much of a
problem.  And the ergonomics are great.  I can easily and cheaply share
the same sequence between Rust and Python objects.  Slicing will also be
much easier (and those are needed in MSA, for example).  All in all,
this makes Python sequence type what it should be: a transparent wrapper
to a to a fully Rust-native object.

I also noticed that sequence distance functions converted sequences to
slices right away, so I changed their signatures. This way they can
accept both `Sequence` and `SequenceMut`.