[b3] likelihood: use MSA
Finally.  That turned out to be a big refactor, because I had to tweak a
bunch of things.  I also forgot to commit my refactor of MSA to be
immutable, so that's also here.

I replaced the hash algorithm which turned out to be insanely slow with
the old sorting one.  I realized that I could just sort the indices by
comparing the sites virtually.