[b3] operators: add `SubtreePruneRegraft`
Had to add a few methods to `Tree`.  Converting `Node` to Python objects
is a monumental pain.  I wonder if there's a better way to do that.

I wonder if it'd make sense to allocate all nodes in advance and
immortalize them.  This way I won't need to touch Python at all to
convert `Node` to Python types.