[b3] operators: fix `NodeSlide` Hastings ratio
BEAST X returns the ratio of 0 for subtree slide moves which don't
change the topology.

This breaks the tree, though.  The Hominini branch is now (chimp,
(human, bonobo)) instead of (human, (chimp, bonobo)).

The reason I'm digging into this right now is because I realized that
the MCMC is walking from -400 to -1200 for posterior, which isn't great.

Raising the threshold improves the posterior (which starts to hang
around -400) and the acceptance rates.  But I'm not sure why I would
need to do this.  Node slide dominates the regressions at 95%, but it
also makes up 92% of all accepts, so it's not regressing at a
disproportionate rate.

I'm not sure that the new Hastings ratio is correct for arbitrary
distributions.  `Uniform(0, 1)` also regresses.  But I'm doing a bunch
of transformations, perhaps that's the problem.