[b3] operators: add `SubtreeSlide`
Well, this is just bizarre.  Replacing `NodeSlide` with `SubtreeSlide`
caused the tree topology to get all messed up.  And the lengths are
weird too.  1.0 to the tips and tiny distances between internal nodes.
It would seem that while the operators previously didn't mix very well,
the real problem lies with either the Metropolis-Hastings (or, namely,
with the Hastings ratios) or the likelihood calculator.  Or both D: