[b3] operator: track timing statistics
Also add `print_operator_timings` to `b3.utils` to pretty-print the
performance data.

The results are at microsecond precision, but they have shown me that
all of the operators take pretty much the same time on such a small
tree.  I suspect it'll be more useful for the large simulations.