[b3] operators: fix the `family` helper function
It's a bit maddening how such a small oversight can cause drastic
changes to the MCMC.  Because the old version wasn't picking leaf
`node`s, all `parent` nodes were higher up.  I only noticed this on a 3
leaf node test after several hours because it wouldn't move the non-root
internal.  And, as far as I can tell, fixing `SubtreeSlide` with this
fixes some of the replication issues the primate example was having.

Now I just have to comb through 6 more operators, 3 priors, and I'll be
ready to implement the priors researchers have actually used in the past
decade.

If you are reading this, please send help