[b3] builtin/priors: fix coalescent
So, this is embarrassing.  The first 3 lines is accounting for trees
where not a single tip is at 0.  But the main error stemmed from me
forgetting to divide the binomial by two.

I thought that I had some fundamental problem because my distribution
had bigger variance.  But building a scatter plot between the coalescent
and the tree length showed me that it was the same line, mine was just
tilted.  My line with smaller inclination showed up as a wider
distribution.  Well, at least that's now fixed.