[b3] transitions: fix a bug which was causing full recalculations
I forgot to update the rate, which meant each step would trigger a full
recalculation.  I didn't notice it while working with the apes example
because there the rate was fixed at exactly 1.0.

This also highlights the importance of using NaN as a dummy value for
floats, since it makes any errors immediately apparent.  This wouldn't
have helped here, though, as I was using the `rate` parameter when
calculating distances.