[b3] operators: simplify `DeltaExchange`
This switches it from BEAST2 algorithm to the BEAST one.  The latter
only handles one parameter, which makes it simpler.

I have launched apes with it and encountered an unexpected problem.  Due
to float precision losses the total sum of frequencies can deviate a bit
from 1 in order of 1e-15.  It shouldn't impact substitutions, because
those calculate the normalization coefficient based on the sum of
frequencies.  And the deviation itself is probably small enough to not
show up anywhere.  But it's something to keep in mind.