[b3] parameters: move the implementation to Python
Python is easier to manage and half of the parameters are Python-native
anyways.  `Weights` turned out to be weirdly slow.  I might have to
reimplement them in Rust again.

Aside from that, I only left `Real` around and made it a scalar.  Using
it should become more ergonomic.  A lot of operators which would
realistically never work with multidimensional parameters had a lot of
complications because of it.  And defining throwable `__float__` was
suboptimal.