[b3] likelihood: get rid of `cached_likelihood`
This commit makes the `likelihood` method idempotent.  This should make
the interface much easier to use.  Now the only way to scramble the
internal calculator state is to call `accept`/`reject` after `propose`
without calling `likelihood`.  Actually, now that I think about it, the
CUDA calculators should block copying projections on `copy_projections`,
so the state will go back to normal with the next `accept`/`reject`
call.  This means that one of the most finicky parts of the API is now
finally robust.