[b3] likelihood: tweak the API (again)
I plan to rework the CPU likelihood tohoods. be able to switch out the
matrix and vector implementations.  The main issue right now is that I
don't know how to deal with frequencies.  Ah, as I was writing this, I
realized that I can also add those generics to the substitution trait.

This change rolls in weights into likelihoods.  I initially separated it
out because I thought that some variants of analysis might benefit from
tweaking the weights.  But that's too many parameters.  And even if I
wanted to implement it the future, it would need to be per-site, not
per-pattern.  And weights aggregate sites into patterns.

Finally, I had to add `num_edges` because it turned out I was
calculating it manually when creating transitions.