[logger] Create crate
I tried using `log`, but enabling `kv` and `kv_serde` pulled in a bunch
of dependencies and still forced to to write tons of glue code.  Given
it's pulling `sval` on the backend I suspect we had double conversion:
serde to sval in `log` and sval-like to serde in my crate.

The new design is simpler, works well for my use case, and, most
importantly, really performant.  I have a large buffer and no copies.
The latter did require me to make a bit of a hacky macro.  I suspect
it'll worsen the compilation times by a bit, but I'm willing to take
that hit, because otherwise I can't do exhaustive operator tests, since
I need hundreds of thousands of steps for those.