[benches] b3_default: create and run `MCMC` in the same thread
CUDA doesn't like the threads being moved.  I presume some of the Rust
guarantees of `cudarc` gets subverted by `multiprocessing`.  The current
workaround is to create the `MCMC` object in the same worker which will
run it.  It does give us startup overhead, but I believe it's negligible
compared to general variance.  And I can't account for this startup
penalty when running BEAST, so it's only fair.