[b3] likelihood/cuda: parametrize the kernels
I initially just wanted to get rid of the constant `num_*` arguments,
but then realized it could be used to dynamically configure scaling too.

I also feel that there might be more optimization potential when it
comes to compiling various kernels depending on the shape of the
workload.  But that'll require a very expansive benchmark suite.